Create `IntakeSubsystem` class and move arm constants
