+++ README.md
@@ -84,14 +84,9 @@ Following the WPILib command based structure we have broken our robot up into a 
 
 ## Digital Input Ports
 
-| Port  |            Device            |
-| :---: | :--------------------------: |
-|   1   |   Center Upper Beam Break    |
-|   2   | Lower Podium Side Beam Break |
-|   3   |   Lower Center Beam Break    |
-|   4   |  Lower Amp Side Beam Break   |
-|   5   |  Upper Amp Side Beam Break   |
-|   6   | Upper Podium Side Beam Break |
+| Port  |               Device             |
+| :---: | :------------------------------: |
+|   1   |   Lower Elevator Limit Switch    |
 
 ## Network Map
